+++ algorithm/pid.h
@@ -18,12 +18,12 @@ typedef struct PID {
     float intergral_limit;
     float output_limit;
 
-    int16_t output;
-    int16_t last_error;
-    int16_t intergral;
+    float output;
+    float last_error;
+    float intergral;
 
 }PID_t;
 
-void PID_Init(PID_t *pid,float Kp,float Ki,float Kd,float intergral_limit,int16_t output_Limit);
-int16_t PID_Caculate(PID_t *pid, int16_t target, int16_t get);
+void PID_Init(PID_t *pid,float Kp,float Ki,float Kd,float intergral_limit,float output_Limit);
+float PID_Caculate(PID_t *pid, float target, float get);
 #endif //PID_H
